requirements.txt
django
gunicorn
Pillow
whitenoise
django_heroku